requirements.txt
pytest
urllib3
requests
pydantic
jsonschema~=4.25.1
python-dotenv~=1.1.1